+++ processor/src/test/resources/input/com/example/autorun/ScheduleDeferredModel.java
@@ -1,0 +1,13 @@
+package com.example.autorun;
+
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.Autorun;
+
+@ArezComponent( deferSchedule = true )
+public class ScheduleDeferredModel
+{
+  @Autorun
+  public void doStuff()
+  {
+  }
+}